Favour single property dispatch over global
See recent commits. This can lead to many fewer renders and faster app
performance.

This may also be necessary for reintroduction in to the quiz cycle
shortly where we have just introduced a gap that didn't exist before;
since it is no longer flushing, there is no reason for the poked cells
to reset their colour to white when the new harpstrata is set. Flushing
was our solution to it before. Removing it caused this problem, but
reintroducing it should fix it again.
